# chore: Remove `Get()` in offerings in favor of `Compatible()`