@jonathan-innis 2024-06-10
> This seems like it would make scheduling slower

Big-O of this change is the same, but I agree that we can probably do better by short-circuiting. I changed this to `HasCompatible()` after discussing a bit more.