@jonathan-innis 2024-06-10
Here's the new Scheduling benchmark results

### Before PR

```
=== RUN   TestSchedulingProfile
scheduled 10110 against 32 nodes in total in 90.314545ms with minValues included false 111942.10190617692 pods/sec
400 instances 10 pods    2 nodes  140.337µs per scheduling    14.033µs per pod
400 instances 100 pods   5 nodes  433.869µs per scheduling    4.338µs per pod
400 instances 500 pods   5 nodes  1.975417ms per scheduling   3.95µs per pod
400 instances 1000 pods  5 nodes  3.58416ms per scheduling    3.584µs per pod
400 instances 1500 pods  5 nodes  4.964902ms per scheduling   3.309µs per pod
400 instances 2000 pods  5 nodes  11.227208ms per scheduling  5.613µs per pod
400 instances 5000 pods  5 nodes  21.250258ms per scheduling  4.25µs per pod
--- PASS: TestSchedulingProfile (14.31s)
PASS
```

### After PR

```
=== RUN   TestSchedulingProfile
scheduled 10110 against 32 nodes in total in 72.144137ms with minValues included false 140136.12776323044 pods/sec
400 instances 10 pods    2 nodes  140.338µs per scheduling    14.033µs per pod
400 instances 100 pods   5 nodes  435.474µs per scheduling    4.354µs per pod
400 instances 500 pods   5 nodes  2.012869ms per scheduling   4.025µs per pod
400 instances 1000 pods  5 nodes  3.596223ms per scheduling   3.596µs per pod
400 instances 1500 pods  5 nodes  4.768862ms per scheduling   3.179µs per pod
400 instances 2000 pods  5 nodes  7.747948ms per scheduling   3.873µs per pod
400 instances 5000 pods  5 nodes  24.973005ms per scheduling  4.994µs per pod
--- PASS: TestSchedulingProfile (12.68s)
PASS
```